pyproject.toml
[project]
name = "function-call-graph"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "GPL-2.0-only"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "networkx (>=3.5,<4.0)",
    "angr (>=9.2.184,<10.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "angr-utils (>=0.5.0,<0.6.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "scipy-stubs[scipy] (>=1.16.3.3,<2.0.0.0)",
    "pyqt6 (>=6.10.2,<7.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
